fixing size of images in a reports (#200)
this fixes:
https://github.com/insightsengineering/teal.reporter/issues/149

Issue:
Large plot like forestplot plot form efficacy app was not accomodated in
given space as dimention were not passed correctly and this found in all
report types as comment
https://github.com/insightsengineering/teal.reporter/issues/149#issuecomment-1761548639.

Solution: 
function `teal.widgets::plot_with_settings`, the function returns a`
list (dim, ...)` that is null if the user hasn't manually set the plot
width and height by clicking the expand button on plot. This is in
contrast to the default values passed to this function. To address this
issue, setting the plot dimensions to `default values `if `null `is
found.

result:

[input_20231016153135944.docx](https://github.com/insightsengineering/teal.widgets/files/12915135/input_20231016153135944.docx)

[input_20231016153128744.pptx](https://github.com/insightsengineering/teal.widgets/files/12915138/input_20231016153128744.pptx)

[input_20231016153122085.pdf](https://github.com/insightsengineering/teal.widgets/files/12915144/input_20231016153122085.pdf)

<img width="647" alt="image"
src="https://github.com/insightsengineering/teal.widgets/assets/6700955/9b27df02-bfff-4735-9894-a8efcdc85624">

---------

Signed-off-by: kartikeya kirar <kirar.kartikeya1@gmail.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Aleksander Chlebowski <114988527+chlebowa@users.noreply.github.com>