decouple `teal.widgets` from `teal` in plot resizing (#250)
Alternative to #248 
Also fixes #249 
Accompanied by https://github.com/insightsengineering/teal/pull/1261

Changed the way that plot rerendering is triggered to fill the container
element.
Resizing is still done by sending messages to `shiny` with 
```
    Shiny.onInputChange(flex_width, document.getElementById(plot_out).clientWidth);
    Shiny.onInputChange(plot_modal_width, 0.87 * window.innerWidth);
```
However, the messages are not sent in response to events registered with
`$(window).on("resize")`, but rather by a `MutationObserver` object. The
callback was being registered in `teal.widgets` but triggered in `teal`,
by running `$(".teal_primary_col").trigger("resize")`. Now the resizing
behavior is contained to `teal.widgets`.

The same callback was being registered on `shiny:connected` - this is
removed.

Also set `overflow-x` property on the plot container to avoid it
sticking out of the container before resizing kicks in.