several test enhancements (#256)
A few enhancements including:
- consistent usage of `grDevices::pdf(nullfile())` replacing
`grDevices::pdf()` - a follow-up on
https://github.com/insightsengineering/teal.widgets/pull/203
- better catch errors in the test `"type_download_srv_table: pagination,
lpp to small"`
- DRY for `plot_types` variable
- remove `expect_download`. This is not _only_ expectation that the
downloadable link exists but it also downloads the file and run snapshot
testing on it. The problem is that this file is named from sys time on
the moment of test run (e.g.
`tests/testthat/_snaps/app_driver_pws_ui/pws-001-plot_20240805_034140.png`)
which creates a new file every time you run the test (that's why our
integration test fails). Moreover, I believe that this is not the intend
of that unit test to test regression of a dummy plot.

---------

Signed-off-by: Pawel Rucki <12943682+pawelru@users.noreply.github.com>