Update shinytest2 for bslib on widgets (#298)
# Pull Request

Fixes #296

While working on the tests I changed `rvest::html_nodes` to
`rvest::html_element`.
The former was deprecated in rvest 1.0.0 some time ago. 

I also removed a dataset created for an example (that was not used) and
added some comments to an example to make it easier to understand the
output.

I've added shinytest2 tests to check these widgets are visible. 

- [x] `plot_with_settings()`: Added testing scrollbar functionality
- [x] `table_with_settings()`
- [x] `verbatim_popup()`
- [x] `get_dt_rows()`: No previous tests.
- [x] `optionalSelectInput()`/`updateOptionalSelectInput()`: No previous
tests
- [x] `draggable_buckets()`: no scrollbar, no previous tests.
- [x] `nested_closeable_modal()`: No previous tests
- [x]  `optionalSliderInputValMinMax()`: No previous tests.

---------

Signed-off-by: Lluís Revilla <185338939+llrs-roche@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: André Veríssimo <211358+averissimo@users.noreply.github.com>