fix scheduled jobs (#300)
Fixes https://github.com/insightsengineering/teal.widgets/issues/295

# Current fix

All 3 strategies are green (besides max which is fine)


https://github.com/insightsengineering/teal.widgets/actions/runs/15777310323

Solution: we needed to update `testthat` because the old version was
dependent on old `digest` that could no longer be installed in our
setup.


# Previous errors on main

```
Failed to build `digest` dependency
```


https://github.com/insightsengineering/teal.widgets/actions/runs/15659252962/job/44114591501#step:9:458

`digest` is a part of `testthat`.

---------

Signed-off-by: Marcin <133694481+m7pr@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>