Fix broken `draggable_buckets` (#309)
Closes #308 

The subscribe callback now expects an object with the priority after
this PR https://github.com/rstudio/shiny/pull/4211. So, if you update to
`shiny v1.11.1` the `draggable_buckets` do not work (the init works, but
since the subscribe is broken it does not update).

In the future we can think about using
[`sortable`](https://rstudio.github.io/sortable/) package to replace a
huge chunk of custom css/js code.