318: Increase Coverage beyond 80% (#322)
# Pull Request

Fixes #318 

## Changes

I have added checks for the ui components and modules. For the ui
functions, I added input validation, component class and snapshot tests.
For the modules, I added input validation and small unit tests. All
modules already have e2e tests.
The e2e tests were already existing but it seems they do not count for
the coverage as they are skipped.

We have not reach 100% due to soft_deprecated functions and mainly
verbatim_popup_ui which I was not able to increase that much from the
previous state.

---------

Signed-off-by: Oriol Senan  <35930244+osenan@users.noreply.github.com>
Co-authored-by: osenan <oriol@praenoscere.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>